feat: support vclusters
To make vclusters work properly, we need to get the vcluster pod
name/namespace in some places but keep using the host name/namespace in
others. We solve this by passing the vcluster annotations that contain
the name/namespace to the shim and then use the host/vcluster names
where appropriate.